Fix DLL preload for the libdynd default install location.
